Move all crowdfunding code to common-ui.
