+++ src/sidebar-overlay/components/AnnotationContainer.js
@@ -6,8 +6,10 @@ import { remoteFunction } from '../../util/webextensionRPC'
 
 import Annotation from './Annotation'
 import TagHolder from './TagHolder.js'
-import CrowdfundingBox from './CrowdfundingBox'
-import CrowdfundingModal from '../CrowdfundingModal/Overlay'
+import {
+    CrowdfundingBox,
+    CrowdfundingModal,
+} from '../../common-ui/crowdfunding'
 import styles from './Annotation.css'
 import { IndexDropdown } from '../../common-ui/containers'
